requirements.txt
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pyautogui>=0.9.54
pywin32>=306
keyboard>=0.13.5
mss>=9.0.1

# AI 및 머신러닝
torch>=2.0.0
requests>=2.31.0
scikit-learn>=1.3.0
psutil>=5.9.0

# RAG 시스템
sentence-transformers>=2.2.0
sqlite3  # 내장 모듈

# 웹 서핑 및 비서 기능
beautifulsoup4>=4.12.0
aiohttp>=3.8.0
schedule>=1.2.0

# LLM 연동 (선택적)
# ollama-python>=0.1.7  # pip install ollama-python